+++ NINA/Utility/Converters/InverseBooleanConverter.cs
@@ -29,7 +29,7 @@ namespace NINA.Utility.Converters {
 
         public object ConvertBack(object value, Type targetType, object parameter,
             System.Globalization.CultureInfo culture) {
-            throw new NotSupportedException();
+            return !(bool)value;
         }
 
         #endregion IValueConverter Members